Fix warnings in murmur3.c when building for 32-bit.

+++ common/murmur3.c
@@ -67,11 +67,11 @@ static FORCE_INLINE uint64_t fmix64(uint64_t k)
 
 //-----------------------------------------------------------------------------
 FORCE_INLINE static void
-MurmurHash3_x86_32(const void *key, int len, uint32_t seed, void *out)
+MurmurHash3_x86_32(const void *key, size_t len, uint32_t seed, void *out)
 {
     const uint8_t *data = (const uint8_t *)key;
-    const int nblocks = len / 4;
-    int i;
+    const size_t nblocks = len / 4;
+    size_t i;
     uint32_t h1 = seed;
     uint32_t c1 = 0xcc9e2d51;
     uint32_t c2 = 0x1b873593;
@@ -115,12 +115,12 @@ MurmurHash3_x86_32(const void *key, int len, uint32_t seed, void *out)
 
 
 //-----------------------------------------------------------------------------
-FORCE_INLINE static void MurmurHash3_x86_128 (const void *key, const int len,
+FORCE_INLINE static void MurmurHash3_x86_128 (const void *key, const size_t len,
                                               uint32_t seed, void *out)
 {
-    int i;
+    size_t i;
     const uint8_t * data = (const uint8_t*)key;
-    const int nblocks = len / 16;
+    const size_t nblocks = len / 16;
 
     uint32_t h1 = seed;
     uint32_t h2 = seed;
